requirements.txt
fastf1>=3.3.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
diskcache>=5.6.3
python-dotenv>=1.0.0
streamlit>=1.32.0
plotly>=5.18.0
